backend/pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "LangChain-based multi-agent system for job application assistance"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.10",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "openai>=2.11.0",
]
